+++ database/migrations/2024_09_26_174719_create_comments_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->unsignedSmallInteger('post_id')->index();
             $table->text('content');
             $table->unsignedSmallInteger('status')->default(1);
-            $table->unsignedSmallInteger('like');
+            $table->unsignedBigInteger('like');
             $table->unsignedSmallInteger('parent_id');
             $table->timestamps();
         });
